feat: heart 기능 추가 #53

+++ lib/feature/store/favorite_store/bloc/favorite_store_bloc.dart
@@ -6,20 +6,25 @@ import 'package:cafein_flutter/data/repository/heart_repository.dart';
 import 'package:equatable/equatable.dart';
 
 part 'favorite_store_event.dart';
+
 part 'favorite_store_state.dart';
 
 class FavoriteStoreBloc extends Bloc<FavoriteStoreEvent, FavoriteStoreState> {
-  FavoriteStoreBloc({required this.heartRepository}) : super(const FavoriteStoreInitial()) {
+  FavoriteStoreBloc({required this.heartRepository})
+      : super(const FavoriteStoreInitial()) {
     on<FavoriteStoreRequested>(_onFavoriteStoreRequested);
     on<FavoriteStoreClicked>(_onFavoriteStoreClicked);
   }
+
   final HeartRepository heartRepository;
   List<MemberStore> favoriteStores = [];
+  List<bool> heartList = [];
   int favoriteStoreCount = 0;
+
   FutureOr<void> _onFavoriteStoreRequested(
-      FavoriteStoreRequested event,
-      Emitter<FavoriteStoreState> emit,
-      ) async {
+    FavoriteStoreRequested event,
+    Emitter<FavoriteStoreState> emit,
+  ) async {
     emit(const FavoriteStoreLoading());
     try {
       final storeResponse = await heartRepository.getMyStores();
@@ -27,7 +32,11 @@ class FavoriteStoreBloc extends Bloc<FavoriteStoreEvent, FavoriteStoreState> {
       final storeCount = storeResponse.data.storeCnt;
       favoriteStores = stores;
       favoriteStoreCount = storeCount;
-      emit(FavoriteStoreLoaded(stores: [...stores], storeCount: storeCount));
+      heartList = List<bool>.filled(storeCount, true);
+      emit(FavoriteStoreLoaded(
+          stores: [...stores],
+          storeCount: storeCount,
+          heartList: [...heartList]));
     } catch (e) {
       emit(FavoriteStoreError(
         error: e,
@@ -35,16 +44,24 @@ class FavoriteStoreBloc extends Bloc<FavoriteStoreEvent, FavoriteStoreState> {
       ));
     }
   }
-
 
   FutureOr<void> _onFavoriteStoreClicked(
-      FavoriteStoreClicked event,
-      Emitter<FavoriteStoreState> emit,
-      ) async {
+    FavoriteStoreClicked event,
+    Emitter<FavoriteStoreState> emit,
+  ) async {
     emit(const FavoriteStoreLoading());
     try {
-      heartRepository.deleteHeart(event.clickedStoreId);
-      emit(FavoriteStoreLoaded(stores: [...favoriteStores], storeCount: favoriteStoreCount));
+      if (heartList[event.clickedStoreIndex]) {
+        heartRepository.deleteHeart(event.clickedStoreId);
+        heartList[event.clickedStoreIndex] = false;
+      } else {
+        heartRepository.createHeart(event.clickedStoreId);
+        heartList[event.clickedStoreIndex] = true;
+      }
+      emit(FavoriteStoreLoaded(
+          stores: [...favoriteStores],
+          storeCount: favoriteStoreCount,
+          heartList: [...heartList]));
     } catch (e) {
       emit(FavoriteStoreError(
         error: e,
@@ -52,5 +69,4 @@ class FavoriteStoreBloc extends Bloc<FavoriteStoreEvent, FavoriteStoreState> {
       ));
     }
   }
-
 }
